refactor: 백준 tag dto class 3개를 static으로 한 클래스로 몰아넣어 1개로 줄임
