+++ src/main/java/vision/cotegory/crawler/baekjoon/BaekjoonProblemService.java
@@ -3,12 +3,14 @@ package vision.cotegory.crawler.baekjoon;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import vision.cotegory.entity.*;
+import vision.cotegory.crawler.baekjoon.SolvedAcProblemDto.SolvedAcTagDto;
+import vision.cotegory.crawler.baekjoon.SolvedAcProblemDto.SolvedAcTitleDto;
+import vision.cotegory.entity.Quiz;
+import vision.cotegory.entity.Tag;
+import vision.cotegory.entity.TagGroup;
+import vision.cotegory.entity.TagGroupConst;
 import vision.cotegory.entity.problem.BaekjoonProblem;
 import vision.cotegory.entity.problem.BaekjoonProblemPage;
-import vision.cotegory.crawler.baekjoon.dto.SolvedAcProblemDto;
-import vision.cotegory.crawler.baekjoon.dto.SolvedAcTagDto;
-import vision.cotegory.crawler.baekjoon.dto.SolvedAcTitleDto;
 import vision.cotegory.repository.BaekjoonPageRepository;
 import vision.cotegory.repository.BaekjoonProblemRepository;
 import vision.cotegory.repository.QuizRepository;
