+++ src/main/java/vision/cotegory/crawler/baekjoon/SolvedAcWebClient.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 import vision.cotegory.exception.exception.SolvedApiException;
-import vision.cotegory.crawler.baekjoon.dto.SolvedAcProblemDto;
 
 import java.util.List;
 import java.util.stream.Collectors;
