--- src/main/java/vision/cotegory/crawler/baekjoon/dto/SolvedAcTagDto.java
@@ -1,11 +1,0 @@
-package vision.cotegory.crawler.baekjoon.dto;
-
-import lombok.Data;
-
-@Data
-public class SolvedAcTagDto {
-    private String key;
-    private Integer bojTagId;
-    private Integer problemCount;
-    private Boolean isMeta;
-}
