--- src/main/java/vision/cotegory/crawler/baekjoon/dto/SolvedAcTitleDto.java
@@ -1,11 +1,0 @@
-package vision.cotegory.crawler.baekjoon.dto;
-
-import lombok.Data;
-
-@Data
-public class SolvedAcTitleDto {
-    private String language;
-    private String languageDisplayName;
-    private String title;
-    private Boolean isOriginal;
-}
